ASoC: Move Sigma Studio firmware loader to ASoC
It has been pointed out previously by some, that the firmware subsytem is not
the right place for the sigma studio firmware loader.
Since the sigma dsp is currently only used in audio products. Also we are aiming
for a thighter integration into the ASoC framework in the future, with support
for ALSA controls for firmware parameters and support dynamic power management
as well. So the natural choice for the Sigma studio firmware loader is the ASoC
subsytem.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>

Hand tweaked

Signed-off-by: Steve Conklin <steve@conklinhouse.com>