ASoC: Sigma: Prevent out of bounds memory access
The sigma DSP firmware loader currently does no proper bounds checking for
access behind the actual firmware size. As a result it is possible that a
malformed firmware can cause a out of bounds memory access.

This patch adds checks which ensure that both the action header and the payload
are completely inside the firmware data bounds before processing them.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>