ASoC: Sigma: Add regmap support
Add support for loading the Sigma Studio firmware using regmap. This allows us
to transparently use SPI or I2C on devices which support this.

For now we keep the old I2C support since we have one user of this which is not
straight forward to convert to regmap, due to variable length registers.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Steve Conklin <steve@conklinhouse.com>